pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi[standard]>=0.128.0",
    "pillow>=12.1.0",
    "psycopg2>=2.9.11",
    "pycountry>=24.6.1",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.31",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10",
]
